requirements.txt
langchain==0.3.25
langchain-openai==0.3.12
langchain-community==0.3.21
langchain-core==0.3.84
langgraph==0.3.34
chromadb==0.5.23
langchain-chroma==0.2.2
langchain-text-splitters==0.3.8
streamlit>=1.38.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
openai>=1.40.0
fastmcp>=0.1.0
